move loading/writing of keys
